Check if indexes exist before adding
Since the name of a migration has been changed, it will be copied over
again and then run on rake railties:install:migrations.

Thus, it needs to be ensured that if an index is already present,
add_index is not used to create it again and thus a check using
index_exists? has been added for all the concerned add_index calls.